Seperate custom widgets to own folder for better overview
